Made move result class to inherit setting attributes and other common methods from, and deleted a bit of redundant code
