+++ src/main.py
@@ -42,14 +42,8 @@ for screen in screens:
 
 sm.current = "home"
 
-# screens = [HomeWindow(name="home"), StartWindow(name="startpage"), RefWindow(name="reference"),
-#            MoveSearch(name="movesearch"), PracticeWindow(name="practice")]
 
-
-class MyMainApp(App):  # Inherits from App class
-    # def __init__(self, default_config=None, **kwargs):
-    #     super().__init__(**kwargs)
-    #     self.scrollable_view = ScrollableView()
+class MyMainApp(App):
 
     def build(self):
         return sm
